+++ source/refined-github.ts
@@ -223,4 +223,4 @@ import './features/pr-notification-link.js';
 import './features/click-outside-modal.js';
 import './features/comment-excess.js';
 import './features/linkify-line-numbers.js';
-import './features/search-page-title.js';
+import './features/search-query-title.js';
